Custom Radar, Vetos, Panel Styling
This version adds a couple of very important options to the panel. First being that there is now a custom radar that uses the open source radar.. To use this feature, you must run the radar.bat file along with the hud.bat file. The second important feature added is the vetos system. Depending on which best of type you choose, map cards will appear for you to select maps, teams, and winners. This is the first attempt at all of this so there will probably be bugs. Please report any you see.